requirements.txt
llama-cpp-python
requests
serde
tqdm
